Add base settings to test UI
