--- apps/sasquatch-kotlin/src/main/java/com/microsoft/appcenter/sasquatch_kotlin/AnalyticsFragment.kt
@@ -1,18 +1,0 @@
-package com.microsoft.appcenter.sasquatch_kotlin
-
-import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-class AnalyticsFragment : Fragment() {
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_analytics, container, false)
-    }
-
-
-}
