+++ apps/sasquatch-kotlin/src/main/java/com/microsoft/appcenter/sasquatch_kotlin/modules/AnalyticsModulePage.kt
@@ -1,0 +1,13 @@
+package com.microsoft.appcenter.sasquatch_kotlin.modules
+
+import android.os.Bundle
+import android.support.v4.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.microsoft.appcenter.sasquatch_kotlin.R
+
+class AnalyticsModulePage : BaseModulePage() {
+    override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
+    }
+}
