+++ apps/sasquatch-kotlin/src/main/java/com/microsoft/appcenter/sasquatch_kotlin/modules/BaseModulePage.kt
@@ -1,0 +1,28 @@
+package com.microsoft.appcenter.sasquatch_kotlin.modules
+
+import android.support.annotation.StringRes
+import android.support.v7.preference.Preference
+import android.support.v7.preference.PreferenceFragmentCompat
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+
+abstract class BaseModulePage : PreferenceFragmentCompat() {
+    protected val uiScope = CoroutineScope(Dispatchers.Main)
+
+    // region Helpers
+
+    protected inline fun <reified T : Preference> updatePreference(@StringRes id: Int, crossinline block: suspend T.() -> Unit) =
+            uiScope.launch {
+                val preference = preferenceManager.findPreference(getString(id)) as T
+                preference.block()
+            }
+
+    protected inline fun <reified T> onPreferenceChange(crossinline block: suspend (value: T) -> Unit) =
+            Preference.OnPreferenceChangeListener { _, newValue ->
+                uiScope.launch { block(newValue as T) }
+                true
+            }
+
+    // endregion
+}
