+++ apps/sasquatch-kotlin/src/main/res/navigation/nav_graph.xml
@@ -7,17 +7,17 @@
 
     <fragment
         android:id="@+id/overviewFragment"
-        android:name="com.microsoft.appcenter.sasquatch_kotlin.OverviewFragment"
+        android:name="com.microsoft.appcenter.sasquatch_kotlin.modules.OverviewModulePage"
         android:label="fragment_overview"
         tools:layout="@layout/fragment_overview" />
     <fragment
         android:id="@+id/analyticsFragment"
-        android:name="com.microsoft.appcenter.sasquatch_kotlin.AnalyticsFragment"
+        android:name="com.microsoft.appcenter.sasquatch_kotlin.modules.AnalyticsModulePage"
         android:label="fragment_analytics"
         tools:layout="@layout/fragment_analytics" />
     <fragment
         android:id="@+id/crashesFragment"
-        android:name="com.microsoft.appcenter.sasquatch_kotlin.CrashesFragment"
+        android:name="com.microsoft.appcenter.sasquatch_kotlin.modules.CrashesModulePage"
         android:label="fragment_crashes"
         tools:layout="@layout/fragment_crashes" />
 </navigation>
